SCRD-2122 Skip config.tgz installation with new inst method
The symlinks which are installed via the config.tgz tarball are not
strictly needed (but nice to have for customers to have the
configuration files in a single place instead of digging through the
different roles/*/templates dirs). So skip the installation when using
the new installation process (set via
"deployer_media_legacy_layout=False").
We might want to create the symlinks later during runtime but for now,
this is the simple solution to make progress with the alternative
installation method.

Change-Id: Iafb4236997f4ba2f89347a921dc5ed63f1df7431