Exclude localhost from supportutils tasks (bsc#1081689)
Use the same host list as for the sos-report related tasks. This should
avoid running the tasks twice (in parallel) on the same node.
"localhost" is basically a duplicate when using "all" in our case. As
the very same machine is also listed by its "real" name in the inventory.
So the install task runs as on "localhost" and e.g. "ardana-ccp-c0-m1" 
(the real hostname) in parallel, causing issues for zypper.

Change-Id: I1c54e482805c6814a7347dbb0a8c57e8d57ed4d7