Make horizon and opsconsole urls available to install UI (bsc#1083381)
Setting "hosts: localhost" for installui-reconfigure.yml locks the run
onto the deployer, but makes HZN_WEB and OPS_WEB vars unavailable to the
install UI.  Setting "hosts: ARD-SVC" makes the vars available, but
attempts to run it across all ARD-SVC hosts where deployer may not be
running.  This fix allows the reconfiguration to only be available
on the deployer while having access to the necessary HZN_WEB and OPS_CON
vars.

Change-Id: I615aac60ed4521d57b5529970f75c49ab7b6262a