SCRD-2848 Add new directory to packager vhost
This patch adds the product media assumed to be under
/srv/www/suse-12.3 for the Ardana venv server to host on port 79.
Ensuring that product media exists in this directory on the deployer
node can be done manually following a prodecure similar to what
we have documented[1] for our older products, but this is also a first
step toward automating that procedure.

An alternative to this was to create a brand new vhost that directly
served the media in /srv/www listening on a new port, but the port we
used in our other product for this, 8091, is already taken for monasca.
Rather than introduce yet another arbitrary port, make the existing HTTP
server listening on port 79 work for this.

In order to ensure the setup-zypp task in osconfig has a consistent
experience for both the deployer and non-deployer node, this patch swaps
the order of the Apache vhost setup so that the repositories are already
available over HTTP before the repo setup happens. This way we don't
have to special-case the deployer node in osconfig-ansible and all the
logic about where the media is actually stored can remain in
ardana-ansible.

[1] https://www.suse.com/documentation/suse-openstack-cloud-7/book_cloud_deploy/data/sec_depl_adm_conf_repos_product.html

Depends-on: Iea38bbe377b1375bd40323eb14f996484b598973
Change-Id: If683a7602d7e7be34d5621239c8856af10669243