SCRD-2848 Auto-mount Cloud ISO to be HTTP served
This patch helps streamline deployment of Ardana when the initial media
is the Cloud ISO. In our older products, we instructed customers to
copy the installer media into a special directory where the deployer
could find it[1]. With this patch, we can automatically mount that media
in the special location where it will be served to the non-deployer
nodes. This works by searching for the Cloud 8 ISO in the zypper
repository list, parsing the iso:/// URI into a local filesystem path,
and mounting that under /srv/www/suse-12.3/x86_64/repos/Cloud which
already has an Apache vhost serving it.

This also adds a become: yes to the check-iso.yml playbook so that if
the ISO happens to be in the root user directory, it can be found.

[1] https://www.suse.com/documentation/suse-openstack-cloud-7/book_cloud_deploy/data/sec_depl_adm_conf_repos_product.html

Change-Id: Ib2cd7f338efb8d7299f887873ecdc73c289876e9