SCRD-2848 Serve local SMT mirrors if available
This patch makes managing subscription channels easier if the user has
already deployed a local SMT mirror on the deployer node. A task is
added to the deployer node to check if there is any content under
known SMT paths and serve those to non-deployer nodes if available. It
does not interfere of those repositories are already present for some
other reason, such as having already been NFS mounted from some other
server. This depends on repo data stored in osconfig-ansible.

Depends-on: Ib623b12c9dad75725a5dbe168148f39d0897fd06
Change-Id: I2f8364fab57959364f445f3c29ac8b56aa6010d2