SCRD-2502 Fix ISO URI detection
Without this patch, the find_zypper_iso module assumes that the Cloud
ISO was installed by using it as a source for an Add-On product in Yast.
It's also possible to install the ISO as a zypper repository using
zypper directly, in which case the URI will not be percent-encoded and
will not have the directory encoded separately. With this patch, both of
the following forms of URIs should be detectable:

 iso:///?iso=SUSE-OPENSTACK-CLOUD-8-x86_64-Media1.iso&url=dir%3A%2Froot

 iso:///?iso=/root/SUSE-OPENSTACK-CLOUD-8-x86_64-Media1.iso

Change-Id: Iafa6c08c55fc423ca302a01ae4cbcaf917a5a5da