Set 'ptf_signing_key' fact during deployer-init run (bsc#1089973)
When deployer-init is called, the 'ptf_signing_key' fact needs to be
available. Otherwise, zypper doesn't trust the PTF repo and the
installation aborts with failures like:

[PTF|http://localhost:79/SUSE-12-3/x86_64/repos/PTF] Valid metadata \
  not found at specified URL
Please check if the URIs defined for this repository are pointing to \
  a valid repository.

Trying todo a "zypper ref" on the failed deployer node shows that the
PTF key needs to be imported into the trust DB first.

Change-Id: Ibc96a9a9792ade4812f47282c69b4659761c8e43