SCRD-4432 Add support for Cloud 9 media
Update the media detection logic to handle both
Cloud 8 and Cloud 9 in a seamless manner which
does not require branching.

This change also creates a new role, gathering all
the product media and repositories information from
ardana-ansible and osconfig-ansible in one place.
This role is then included by top-level playbooks
wherever this information is required.

Change-Id: Idb787a688b3bf8951846261ba38507d5c53b5e19