SCRD-5330 Freezer not available in Cloud9/Rocky
Freezer is no longer being actively maintained upstream so there is
no stable/rocky release. As such support for it is being dropped from
the Cloud9 release. Therefore we need to remove references to freezer
from the ardana-ansible repo, but also in the existing playbooks and
the list files being used to dynamically generate the ardana-*.yml
top-level playbooks.

Change-Id: Id7423ff0e1127afd4f80eaeba6b69fd51c60e75a