SCRD-5480 Re-work product version management
The path that is used to specify where the venv's are searched
for is partly derived from the ardana_verion.yml file which is
dynamically generated during the RPM build process, mirroring
the legacy HOS dynamic path setup mechanism.

However the reasons for dynamically generating this file in HOS
don't carry over to Ardana so we can manage the version setting
within the ardana-ansible repo itself on the master branch, which
is what the new ardana-version role is for.

For now have disabled the occurrences of including values from
ardana_version.yml with the check for being a legacy deployer
setup where appropriate, and include the ardana-version role
either explicitly in role lists or implicitly via dependencies
in appropriate role meta files. The disabled includes can be
removed when we cleanup the legacy style deployment code in a
subsequent patch.

Note that the current set of venv RPMs in Cloud9 are still using
the incorrect ardana-8, rather than the correct ardana-9, path
for their install path. This patchset includes a workaround to
deal with that, by ensuring that ardana-9 is created as a link
to the ardana-8 directory if only the old version exists. This
workaround can be removed once the changes to update the venv
RPMs install path have landed.

Add git ignorance of vi swap files.

Change-Id: Ieece3c3251a0a46d55c862f5f824bbb0b33b2709