SCRD-5576 Merge ardana-init.bash into ardana-ansible
Issues: SCRD-5576, SCRD-5605

Currently the ardana-init.bash script is pulled into the ardana-ansible
RPM from the ardana/ardana.git repo as part of the RPM build process.

Instead we would like to include the ardana-init.bash script directly
in the ardana/ardana-ansible.git repo meaning that we can actually
relate the SHA of the ardana-ansible RPM to the ardana-init.bash
version that is being used.

This is the first part of the set of changes needed to implement this
change, by adding the ardana-init.bash script to this repo under the
roles/deployer-setup/files area. Additional changes will be needed
to the RPM build process to change where it will retrieve the script
from, once this patch lands.

Change-Id: Ie039d35dca97ce95577de552d2c8cdf37279fd33