SCRD-5576 Remove redundant legacy deployment actions
We no longer support the legacy deployment workflow for Ardana so
remove the lingering actions related to that workflow, and the now
unnecessary when conditions for the SOC/CLM deployment workflow.

NOTE: We need to retain the deployer_media_legacy_layout definition
in the hosts/localhost until all other repos have been updated to
remove their use of that fact.

Change-Id: Ic3c48178e2f1bb66fdbb5b03c3dad62e5911d415