SCRD-7376 Eliminate duplicated action output
Issues: SCRD-7376

Some ansible actions, such as shell or command, were having their
output duplicated because we were explicitly specifying the path
for the ansible installed site area's callback_plugins, and either
by design, or due a bug, the ansible that we are using, 1.9.6, is
running both the callback_plugins from it's install area and any
specified by the callback_plugins config setting, meaning that we
run the human_log plugin twice for the actions that it integrates
with.

This patch changes our generated ansible.cfg files to not include
the ansible site area paths, but rather use the same path as used
by the /etc/ansible/ansible.cfg default configuration file. Note
that this path doesn't exist in our default setup, so should have
no effect on the operation. This change is being applied to all
possible plugin type config settings, not just callback_plugins.

Change-Id: Ie2c165ba66c00aa48c41ceac7cc00804b544c74b