Improve user experience (bsc#1114632)
This patch fixes the following:
  Introduces more intuitive playbook parameters
  Deprecates confusing status playbook parameters
  Makes pending CLM update message more prominent
  Adds node name and None status as necessary to brief message

Customer requested that the confusing set of playbook parameters
update_status_var, update_status_reset, and update_status_set
resulting in 16 different combinations of values be replaced with a
more intuitive and smaller set of parameters. This patch introduces a
more intuitive set of 3 boolean parameters: pending_clm_update,
pending_system_reboot, and pending_service_update. These variable
names were in use and had to be renamed. The existing, internal
variables were prepended with "has_". The older set of parameters
were not removed, only deprecated, but may be removed at a later
date.

This patch moves the command to turn off the pending CLM update flag
higher in the warning message to make it more prominent. It is a
highly recommended step.

The node name is missing and the output is blank in the brief status
if there are no pending actions. This patch adds the node name and
"None" to make the message more clear.

Change-Id: Ie39131329ab11c1b263491e7e45ac4b94a32738a