SCRD-5899 Safely update .ssh/known_hosts (SCRD-5899)
The previous implementation of the ardana-ssh-keyscan role was using a
shell append operation to update the .ssh/known_hosts, but since the
operation is run for each node in the play, but delegated to localhost,
it means that all actions run on the deployer node, and if you have a
large number of nodes in your cloud then there is a risk that you could
end up with parallel shall actions running at the same time, which may
corrupt/overwrite each others updates to the known_hosts file.

A better approach is to collate all of the keyscan results into a single
ansible fact and handle all of the updating inside a single shell
action that acts upon the collated data.

Change-Id: Ief999bc89a5b76d83f28dd29eee4a77b3aa64b9f