Adds repository list parameter (bsc#1122825)
Some customers are very particular when it comes to updating and/or
patching their production systems. These customers would like to have
as much control as possible of the update playbooks. This fix adds a
repository list parameter to the update-pkgs playbook and gives them
that finer control they are requesting.

Change-Id: I1978422fe27d012639fa59ac51f8241ad32441ca