Ensure Cloud8/SLE-12-SP3 repps still served (SOC-9840)
When the Cloud9 version of ardana-init is run it updates the vhosts
definition used to enable service the venvs and local repos via port
79 to only serve the content under /srv/www/suse-12.4. As a result
any remaining Cloud8/SLE-12-SP3 based nodes, within the cloud, that
depends on the deployer to serve their repos, will no longer be able
to access them.

This patch modifies the Cloud9 deployer initialisation code to ensure
that all existing release hierarchies are configured to be served on
port 79.

A new deployer-init-apache role has been created to handle this work,
which leverages the ansible meta action's flush_handlers mechanism to
avoid the need to split the install/configure and start into separate
top level plays.

A new top level playbook, deployer-init-apache.yml, has been created
to invoke this new role, and included at the start of the existing
deployer-init.yml; this new playbook can be run at the end of the
upgrade process once the Cloud8/SLE-12-SP3 repos are no longer needed
and have been removed.

Change-Id: I62ee0f558cb8039de43937394282c869a7918a77