Enable Cloud9 parallelised upgrade workflow (SOC-10484)
Modify the order of roles/deployer-setup/files/upgrade.list to ensure
that the FND-CLU-upgrade.yml playbook is run early enough to permit
Cloud 8to9 upgrading.

Also include the galera-upgrade entry so that we support upgrading the
MariaDB/Galera cluster correctly as part of the upgrade. This will also
ensuire that the DB services get (re)started appropriately.

Modify the roles/deployer-setup/templates/ardana-upgrade.yml.j2 to
enable it to be used for Cloud9 upgrades.

Change-Id: Id39f93a8480a5eb337f28bc86895982204428dbd