Ensure tls-upgrade.yml called by ardana-update.yml (SOC-9942)
Currently the dynamically generated ardana-update.yml doesn't include
tls-upgrade so we won't invoke tls-upgrade.yml as part of the update
process. This hasn't been an issue up to now as the tls-ansible code
base hasn't changed since prior to the Cloud 8 release. But this change
will be needed if we want to enable customers to be able to correctly
update between incremental MU updates that include tls-ansible changes.

Change-Id: I0279558e7529cbff42eded3d864c5cf8194b8b35