Ensure rabbitmq-server started after packages updated (SOC-11070)
During the zypper dist-upgrade that is run on all nodes in the cluster
to upgrade the nodes packages from SP3/Cloud8 to SP4/Cloud9, sometimes
the automatically triggered restart of rabbitmq-server fails due to
the dependent epmd symlink not being valid (because the associated
erlang package is being updated at the same time; triggering a start
of the rabbitmq-server servics on the relevant nodes after the update
completes ensures the services are running before we start the upgrade
process for the service configuration, thus avoiding secondary issues.

This may also affect a zypper patch or update action within a release
stream if both the rabbitmq-server and erlang packages have newer
versions to be installed.

As part of this I have created two new top-level sub-playbooks for the
upgrade workflow that are invoked during the ardana-upgrade.yml plays
before and after the ardana-update-pkgs.yml playbook:
  * ardana-pre-update-pkgs.yml
  * ardana-post-update-pkgs.yml

I have moved the invocation of the manila-pre-upgrade.yml into the
ardana-pre-update-pkgs.yml, and the actions being taken there are
appropriate for that phase of the upgrade workflow.

In ardana-post-update-pkgs.yml have added an invocation of the
rabbitmq-start.yml playbook, ensuring that rabbitmq-server is being
started everywhere before proceeding with the upgrade workflow.

Change-Id: I14b1a0f479a2d3d9292f0fc94cb3a1a5ed033765