Serialise the _ardana-update-base.yml zypper actions (SOC-11083)
In some cases we are seeing that even just ensuring that services
that failed to restart during the zypper dist-upgrade action are
started again afterwards before proceeding with the rest of the
ardana-upgrade.yml playbook, is not enough to guarantee that
RabbitMQ or MariabDB/Galera restarts reliably.

Serialising the play in _ardana-update-base.yml that runs the zypper
action on the non-deployer nodes, in addition to ensuring that certain
failed services are restarted afterwards seems to work more reliably,
though further refinement of this workaround may be needed.

Also forcing a flush of page cache after upgrading a node using zypper
dist-upgrade seems to release cached pages that otherwise stick around
until we reboot should help eliminate memory pressure issues.

Change-Id: If4157108db8b3e12a14588695c2cf547428cc1e7