Update the Swift XFS inode size check (SOC-10300)
The smaller XFS inode size doesn't actually pose a problem for
upgrading a SOC 8 cloud to SOC 9. It just means that there will
not be enough space avaulable in the XFS inodes to permit certain
XFS data integrity features to be enabled, that are recommended
defaults when formating an XFS file under SLE 12 SP4.

Update the pre-upgrade validation checks to reflect this better
understanding of the implications of the smaller inode size, and
update the reported message to be reported as a NOTE rather than
a failure.

Also find and report on all Swift XFS file systems, not just the
first one on each Swift node.

Also fix all relevant plays to run against the MON-API rather than
KEY-API host groups.

Change-Id: I6813aeaeca5dee4b4480eac1971a8a1cff3ed43d